Fix issues when both base url and subdomains are used
Before this patch, we simply used string.replace() to update urls in the
file. However, this is problematic in cases where we link to both a base
url and subdomains in the same document - e.g. when we link to both
github.com itself and to github.com/some/reop.

To fix this, we now use regex and only replace URLs if they are
preceeded and followed by braces(because Markdown links are formatted as
[link](url)).